[refactor] #28 : 컴포넌트 분리
