+++ src/feature/main/ui/index.ts
@@ -1,2 +1,3 @@
 export * from "./FileDndBox";
 export * from "./FileUploadTable";
+export * from "./PageHeader";
